Feature/event handler subsection
<img width="1143" alt="Screen Shot 2020-05-28 at 16 58 29" src="https://user-images.githubusercontent.com/4676536/83205858-81edae00-a104-11ea-9d60-46b4c707e2b1.png">

just reads a list of all the event handlers and displays them, no CRUD. I think next course of action would be to enable users to click through to edit the handlers, and the ability to unset them.

**Hyperlinking + More metadata**

Event handlers are not representable in a format amenable to the inspector so users will need to read the the code, its our job to get them to the relevant place quickly. Second, our present jsx parsing is pretty basic and needs to support the many forms an event handler definition could take. For that, we need a discussion of how to expand our JSXAttribute parsing. My suggestion is that we need to be able to handle the following scenario. n.b. `EventHandlerX` should be understood to stand in for `mouseDown` etc…:

```jsx
import { importedFn, fnObject } from './importedFn'

const arrow = () => 'arrow'

function ancientFunction() {
  return 'fn'
}

...

      <View
        eventHandler1={() => 'hello'}
        eventHandler2={function () { return 'world' }}
        eventHandler3={ancientFunction}
        eventHandler4={arrow}
        eventHandler5={importedFn}
        eventHandler6={fnObject.memberFn}
        eventHandler7={forbiddenFunction} // n.b. forbiddenFunction is not defined
      />
```

1. eventHandler1: parse the text of the function, and the type being an arrow function expression
2. eventHandler2: parse the text of the function, and the type being an anonymous function expression
3. eventHandler3: parse the name of the variable, and the location of the definition
4. eventHandler4: parse the name of the variable, and the location of the definition
5. eventHandler5: parse the name of the variable, and the imported location of the definition
6. eventHandler6: parse the name of the variable, and the location of the definition
7. eventHandler7: parse the name of the variable, and that there is an error with this variable name

Presently all of these are parsed as "ATTRIBUTE_OTHER_JAVASCRIPT", which only gives us some basics, as an example:

```
{
  type: 'ATTRIBUTE_OTHER_JAVASCRIPT',
  javascript: "() => 'hello'",
  transpiledJavascript: "return (function () {\n  return 'hello';\n});",
  definedElsewhere: [],
  sourceMap: {
    version: 3,
    sources: ['/src/app.ui.js'],
    names: ['(', ')', 'hello'],
    mappings: 'OAcsBA,CAACC;AAAAA,SAAMC,OAAND;AAAAA,CAADD',
    file: '/src/app.ui.js',
    sourcesContent: [
      ...
    ],
  },
  uniqueID: '7958dd9c-f594-496f-a9a7-281bc0d7620d',
}
```

I believe this means we need to have more types of JSXAttributes, namely:

1. arrow function expression
2. function expression
3. variable
4. ???