dont run domWalker on scroll (#5430)
**Problem:**
When we scroll the canvas, we always trigger the `domWalker`, since the check for `nothingChanged` always returns false (even though it's just the canvas coordinates)
<img width="512" alt="image" src="https://github.com/concrete-utopia/utopia/assets/7003853/e1db20b4-902d-4a8f-9b86-8e0838625e34">


**Fix:**
This is the first fix for this. The simplest fix is to have a special case for `SCROLL_CANVAS` action - where the `domWalker` doesn't need to run. 

** Note: **
A more complex fix can be to have the actions themselves state whether they need the `domWalker` to run or not - or have another flag in addition to `nothingChanged` that checks only for `domWalker` relevant changes.

**Manual Tests:**
I hereby swear that:

- [X] I opened a hydrogen project and it loaded
- [X] I could navigate to various routes in Preview mode

Fixes #5385