VariablesInScope Fix: for component arbitrary blocks, change insertionCeiling to rootElementPath (#5818)
This is a preparation PR for #5815

**Problem:**
The ComponentRendererComponent saves variables declared in its arbitrary
block into `VariablesInScope`. However, the `insertionCeiling` is:
1. not guaranteed to be non-null, even though a null value would not be
sensible here
2. points to the _element instance_ of the component instead of the
component's root element. since the insertionCeiling is a ceiling above
which insertion is not allowed, it must be the component's root element.

**Commit Details:** (< vv pls delete this section if's not relevant)

- Adding null check before pushing to VariablesInScope
- Use `insertionCeiling: rootElementPath`
- As @bkrmendy pointed out, it actually makes more sense to make
insertionCeiling non-nullable, so I did that and fixed the fallout

**Manual Tests:**
I hereby swear that:

- [x] I opened a hydrogen project and it loaded
- [x] I could navigate to various routes in Preview mode