unified cartouche datatype (#5836)
**Problem:**
The cartouche sometimes shows `[ ]` icons, sometimes the `{ }` icon, and
other times the 🌺 icon.
But these icons were not synchronized across the various places
rendering cartouches. Despite them sharing the same component they don't
share the same data sources, so some more unification was needed.

**Fix:**
Create a new helper function `getCartoucheDataTypeForExpression` which
takes a unified approach to coming up with the right icon to show for
the cartouche, so it is the same inside the data picker, navigator, and
the inspector.

**Commit Details:**
- Created `getCartoucheDataTypeForExpression`
- Created helper `getSpiedValueForIdentifierOrAccess` – I tried to look
for code to reuse for this one, but couldn't find much
- Added `datatype: CartoucheDataType` to the various cartouche wrapper
component props
- Filling `props.datatype` with data from
`getCartoucheDataTypeForExpression`

**Note:**
- The data picker is the last remaining place which uses a separate data
source called `childTypeToCartoucheDataType` that is coming from the
`VariableInfo` internal type. the reason I'm keeping it separate is
because for dozens of cartouches I'm worried it would add extra cost to
call out to `getCartoucheDataTypeForExpression`. the results today match
up, but if there is a slight divergence we can switch this to
getCartoucheDataTypeForExpression too, but it needs some careful
performance investigation then.