Data Picker Simpler Property Accessors  (#5952)
**Problem:**
The data picker would generate code such as
`reviews['references']['nodes'][0]['id']` which is hard to read and does
not pass our bar of trying to write code a real developer would write.

**Fix:**
Using a simple regex from here
https://stackoverflow.com/questions/1661197/what-characters-are-valid-for-javascript-variable-names,
we test if the property accessor name is trivially safe. if it is safe,
we append it with a `.id`. if it's not safe, we append it as `['id']`.
We keep array element accesses as number.

**Result:**
`reviews.references.nodes[0].id` – this is how I would write it too!