Reparenting out from a scene should have low fitness (#5950)
**Problem:**
Reparenting out from a scene is risky. If your component/element depends
on anything from the context (e.g. a remix loader), then reparenting it
out from the scene will certainly cause issues.
See e.g. this issue:
https://github.com/concrete-utopia/utopia/issues/5915
The worst part is the an exception happens even during mid-interaction,
so you don't even have to drop the element to the storyboard, it is
enough to temporarily drag it over that.

**Fix:**
The best would be to detect when the dragged elements depend on the
scope and cause problems when dragged into a different context. However,
until then, as a heuristic, it seems useful to decrease the fitness of
the reparent strategy when you drag the element out from its scene. So
the move and reorder strategies will be the default, and you really have
to manually select the reparent strategy to force it. This also fixes
the problem of an automatic mid-interaction crash.

**Manual Tests:**
I hereby swear that:

- [x] I opened a hydrogen project and it loaded
- [x] I could navigate to various routes in Preview mode

Fixes https://github.com/concrete-utopia/utopia/issues/5915

---------

Co-authored-by: Balazs Bajorics <2226774+balazsbajorics@users.noreply.github.com>