Fix/wrap undo (#5990)
**Problem:**
When wrapping an element and then immediately undoing, the editor gets
in a stuck state showing a "no entry" mouse pointer.

**Investigation**
The issue turned out to be a problem in the `AddElements` command
executed by `insertIntoWrapper` called by `UPDATE_FNS.WRAP_IN_ELEMENT`:

The AddElements command sets _the unpatched_ EditorState's
`canvas.controls.reparentedToPaths` to an array containing the added
element. This problem was then compounded by the undo's
`restoreEditorState` setting `EditorState.canvas.controls` to
`currentEditor.canvas.controls` which means we always keep the current
reparentedToPaths alive, even if we undo!
This normally shouldn't be a problem because we have an implicit,
_unchecked_ contract that
`UnpatchedEditor.canvas.controls.reparentedToPaths` is always an empty
array.
(As @bkrmendy pointed out, we should have stronger types in place that
disallow setting the unpatched editor state to a non-empty array here.)

**Fix**
- The `AddElements` command now uses `AddToReparentedToPaths` command
instead of hand rolling the setting of `reparentedToPaths`
- Both `AddElements` and `AddToReparentedToPaths` become interaction
lifecycle aware: `AddToReparentedToPaths` only adds to the reparented
paths `mid-interaction`, meaning it will NEVER emit a non-empty array
for the unpatched editor.