Full length UIDs (#5986)
Part of #5654

**Problem:**

> The UIDs are too short

…and therefore prone to collisions.

**Fix:**

> I made them longer

…so they won't collide.

This PR adds support for full-length UIDs. This is done with a split
approach between production and test environments, with the theoretical
ability to configure the desired UID length (but realistically they'll
just stay as they are - 3 chars for tests, 32 chars for prod).

When generating a consistent UID (remember! we'll want to get rid of the
existing UID checks in the future, so we can parallelize the parsing):
- for test environments reuse a good chunk of the previous logic,
generating 3-chars long UIDs that are either a portion of the original
or an incremental alphabetical string (which is now done as an actual
increment rather than doing nested loops which would also be a nightmare
to parametrize).
- for production environments, use either the verbatim hash (128 bit, 32
chars) or a random UUID v4 (with stripped dashes).

**Manual Tests:**
I hereby swear that:

- [x] I opened a hydrogen project and it loaded
- [x] I could navigate to various routes in Preview mode