Navigate to the annotation in the sidecar file (#5999)
**Problem:**
Followup to https://github.com/concrete-utopia/utopia/pull/5989
Clicking on the Components inspector section's ◇ logo opens the
annotation file for the given component AND jumps to the annotation of
the component in the file.

(Second task of https://github.com/concrete-utopia/utopia/issues/5971 )


**Fix:**
I used the Babel parser to find the component annotation in the sidecar
file.
- It can find the annotation if it in the exported object literal, or
even if it is in a variable referred from this file. It does not support
imports though, everything has to be in this file.
- This is a best effort to find the annotation, if it can not do it like
this (e.g. the annotation is dynamically generated, or anything which
the parser is not prepared for), then it just doesn't save the bounds.
Clicking on the button will still open the sidecar file.

The bounds are stored in the component descriptor source object (which
until now only stored the file path itself).

**Manual Tests:**
I hereby swear that:

- [x] I opened a hydrogen project and it loaded
- [x] I could navigate to various routes in Preview mode