Roll your own tab (mostly empty) (#6030)
This is groundwork for
https://github.com/concrete-utopia/utopia/pull/6027.

The `Roll Your Own` floating thing made in the spike becomes a tab in
the right pane (in the future we might want to have it as a floating
panel, but we can incrementally extract it later on).

The `Grid` section is left deliberately empty so we can fill it in with
the other grid-related PRs.

**Context/history**

As part of the exploratory work on grid interactions
(https://github.com/concrete-utopia/utopia/pull/6027) I added a way to
test individual interaction/UI pieces in the spirit of [the classic
Steve Jobs calculator
story](https://www.reddit.com/r/mac/comments/o89l6i/the_apple_team_created_an_app_so_that_steve_jobs/)
- called `Roll Your Own Grid`

<img width="368" alt="Screenshot 2024-07-01 at 17 55 30"
src="https://github.com/concrete-utopia/utopia/assets/1081051/c99a46a8-60ba-4a7c-ae03-ed4145ea7c3a">

There's a good chance this stuff will be useful for more than Grid in
the future (e.g. the inspector, the canvas etc) so we can reuse this
idea in other contexts, while keeping the original concept.

**Note**
The PR includes a demo feature type for `Grid` for illustration purposes
only; it should be replaced with the actual features.