grid-specific hug strategy (#6087)
## Problem
The basic hug contents strategy doesn't really make sense for grids.

## Fix
Create a hug contents strategy tailored to grids. When setting a grid to
hug along an axis, if the grid doesn't use any `fr` units in
`gridTemplate{Row | Column}` along that axis, the width/height prop is
removed from that axis, so the grid is sized by its template rows or
columns. If there's an `fr` unit in`gridTemplate{Row | Column}`, the
strategy is not applicable because the `fr` only works if the grid is
sized